Only show file upload form when form is selected
Turns out there's some confusion on how to use the application when
the file upload is visible.  As such we're hiding the field.

Note that this requires our current setup of dynamic forms, in which
the top-level form specifies a besluitType.  We could move this input
into a custom component which is conditionally rendered.