Updates for new dynamic forms
New dynamic forms allow you to specify certain conditions on forms to
hold.  Eg: when a form is ready to be sent.  The new dynamic forms are
a work in progress, these are the changes necessary to render the new
forms.  Further commits should ensure we consume the states in order
to show or hide buttons correctly.