Always show "Bekijk" in conversation and share model vs id
The Bekijk did not always show up.  The design does include this
feature in the wide view.

The object that was shared was shared by id, rather than the model.
When a primitive value is given to a route, the moedl hook will always
run, slowing down the application.  By sharing the 'row' model, we can
render the objects immediately without passing by the router.  Thereby
rendering the view faster.