Format times using moment.js
This content was formatted beforehand using content in the model.  The
model-rendering methods were removed earlier on.  Correct rendering of
the dates is now handled using moment.

We've also configured moment to use this same method of rendering each
time.  This would help if the designers ask us to render dates in
other places too.