Render tags in template based on what is visible
Rendering content can be served by ember primitives rather than by CSS
classes.  The Ember primitives are easier to recognise and they are
faster too.  This conversion helps in that regard.

Ideally speaking we would expose the display logic towards the outer
components, rather than handling them internally.  This may not be
feasible in the time we currently have hence we may leave it for a
later iteration.