Set laatsteBericht when reaction is created
When a new reaction is created, we should set the 'laatsteBericht'
property.  Although this property could be configured by a reactive
microservice or a reasoner in the future, we currently explicit all of
this state in the services that create and manage the data.  By
setting this value, it shows correctly in the overview.