Simpler url structure
Previously, each utl fieldcosisted of two filed: aprotocol field and a body field. Now, these are combinr=ed to a single field. Also, the data type of the url field is modified in the file-address model. A lot of changes have also been made in the send action on inzending-edit component to make the flow simpler. At the moment of this comment, there is no error, but I didn't have enough time to see what happens exactly, whether the data is saved or not....