Render the url-box component only when necessary
If (not isSent) users will need the component to add new urls or edit the existing ones. If isSent==true then users won't be able to edit it. This time, we check to see whether the toezicht contains any url(/fileAddress). If is doesn't then we don't render the component.