Bugfix/melding view (#2)
* New state: editMode

This indicates whether user shall be able to edit the fields or not

* Remove unnecessary prompt

This calss is used to indicate a mandatory input. Currently, it prints 'VERPLICHT'. This feature is needed only in the edit mode.

* A new route for viewing a toezicht

* Adopt the UI to the new route

* Set editMode on component call

* Update the condition

* Render the url-box only when needed

The url-box is needed when user is creating a melding for the first time or is editing an existing one. However, when user is just viewing a melding, it is only needed if there is something to show, i.e. the fileAddresses array contains at least one item.

* Adopt UI to edit mode

When creating a melding, users need an input field to add a url. When editing a melding, they need the same input field for editing it. But when thay are just viewing a melding, they may expect a simple link to click on.

* Update indentations and remove redundant parens

* await is unnecessary

* Fixed: the 'Voeg link toe' button was not working

This is not an elegant solution, states are now much more complicated than they should be.

* A simpler condition

This component can decide on its own for its state, because it has roughly all the necessary data: isSent.

* Get rid of a redundant prompt

When user can not add any urls, s/he does not need to know what an appropriate url is.

* Get rid of redundant route

Since we have given the inzendingen-table component the authority over its state, we don't need a route to tell viewing/editing states apart.

* New state: editMode

This indicates whether user shall be able to edit the fields or not

* A new route for viewing a toezicht

* Adopt the UI to the new route

* Set editMode on component call

* Adopt UI to edit mode

When creating a melding, users need an input field to add a url. When editing a melding, they need the same input field for editing it. But when thay are just viewing a melding, they may expect a simple link to click on.

* url-line component will take care of each url

* call the url-line component

* Render the url-box component only when necessary

If (not isSent) users will need the component to add new urls or edit the existing ones. If isSent==true then users won't be able to edit it. This time, we check to see whether the toezicht contains any url(/fileAddress). If is doesn't then we don't render the component.

* get call is redundant

* Readability: moved logic to the component

* Give the url-line template full control

* Prevent ember-generated wrapper div

* Render sent urls as links

* Prempt is corrected

* Get rid of unuse property

This might have been left over from a previous merge.

* Get rid of redundant logic

Because we currently only fetch the files when the form is submitted, there's no need for the save.

* Provide a default value

This way, it will be more suggestive for future developers.

* Fixing two typos

* Breaking the code into more readable steps

* No prompt when there is no cached file

This needs further discussion  with the UX designer.

* Get rid of redundant argument

The 'inzending-edit' component already has total control over its state. So, passing this argument is both redundant and error prone. It should actually been removed in a previous commit where the 'editMode' property was removed from the component.

* Render an input field in both cases

This needs further discussion with the UX designer.

* Rearrangement

Having the link to the cached version under the original url is arguably more sensible to the user. Because the  original url is the main player here.