Set bestuurseenheid as leidinggevendenbeheer model
Every subroute of leiddingevendenbeheer retrieves the bestuurseenheid of the logged in user by injecting the currentSession service and getting currentSession.group. Therefore it may be better to set the bestuurseenheid as model of the leidinggevendenbeheer route and access it from the child routes using the modelFor method. It will reduce code duplication in each of the child routes.