Feature/general improvement (#11)
* Set bestuurseenheid as leidinggevendenbeheer model

Every subroute of leiddingevendenbeheer retrieves the bestuurseenheid of the logged in user by injecting the currentSession service and getting currentSession.group. Therefore it may be better to set the bestuurseenheid as model of the leidinggevendenbeheer route and access it from the child routes using the modelFor method. It will reduce code duplication in each of the child routes.

* Remove an unnecessary async

* Remove unused 'showPreferences' attribute

* Get rid of unnecessary model.save()

There is no need to do 'await this.model.save()' since the model hasn't changed.

* Fix a wrong name: email -> telephone

* Rename action 'setPersoon' to 'selectPersoon'

'setPersoon' seems to indicate that a persoon property will be set on the current controller, but the action actually transitions to another route.

* Add a missing '.'

* Remove trivial method 'async save()'

The 'async save()' method doesn't add any value since it's only a one-liner (and it's not consequently used).

* Avoid passing parameters to component

* Remove unnessecary `await`s

* Remove an unused computed property: canConfirm

* Extract functionaris form to a component

* Remove the else branch and always execute it

* Use functionaris-form component in the edit route

* Fix a typo

* Refine contact-info subroute's flow

* Merge two similar actions (addPeriod, bewaar)

* Fix the address of the controller property

* Change contact-info route's model

* Set more clear action names

Currently, we don't show a confirmation dialog. This needs to be discussed with the designer.

* Fix the controller path

* Replace TAB indentation with SPACE

* Generate explicit bestuursfuncties.index route

Use leidinggevendenbeheer.bestuursfuncties.index route to display content that only needs to be shown when no child route is opened instead of using the construction with 'hasActiveChildRoute

* Remov unused computed value 'hasActiveChildRoute'

* Generate explicit index for the functionarissen route

Use leidinggevendenbeheer.functionarissen.index route to display content that only needs to be shown when no child route is opened instead of using the construction with 'hasActiveChildRoute'.

* Get rid of the trick/hack

The hack to refresh the model is not needed anymore since the functionarissen list has now moved to the index route which is not a parent of the current node.

* Add indentation

* Restructure the routes

Create a bestuursfunctie route as the first level child of the leidinggevendenbeheer/bestuursfuncties route and make the contact-info and functinarissen (with all its children) subroutes of it.

* Remove unused import

* Remove unused property

* Conform the code to the data-down, actions-up rule

The bestuursfunctie-card component used to do the route transition which was against the data-down, actions-up pattern. Now, the transition is handled by a method which is passed to the component at initialization time.

* Replace yield with the actual static content.

* Remove an unnecessary ';'

* Remove unnecessary semicolons

* Get rid of redundant 'bewaar' action

* Add some documentation

* Replace html comment with handlebars comment

* Add some documentation

* Get rid of the redundant saveData method

* Extract shared code to a new component

By help of this component, styling of the step indicator text in 3 different pages is controlled from one place.

* Replace html comment with handlebars comment

* Remove unused addPeriod action

* Remove unnecessary await

* Remove unneeded indentations and empy lines

* Remove left over code

This code is moved to the functionaris-creation-step component.

* Disable UI button while saving

* Set isSaving to false by default

* Prevent re-requesting the existing statuses

Updating any bound attribute (like those incorporated in computed values) will trigger the didReceiveAttrs hook. We don't need request backend for the data we already have.

* Use mut helper instead of a dedicated action

* Adjust indentations

* Remove indentations from blank lines

* Remove unnecessary async/await

* Remove redundant legacy condition

The hasActiveChild property has been removed.

* Update route path

The bestuursfunctie_id was duplicated in the url.

* Read bestuursfunctie from parent route

* Move controller manipulations to setupController

* Remove redundant setupController

The controller doesn't need this property since the template of this route is empty.

* Move route and controller logic to the index route

move the model hook of the leidinggevendenbeheer.bestuursfuncties route to the leidinggevendenbeheer.bestuursfuncties.index. The model isn't needed in the leidinggevendenbeheer.bestuursfuncties route. Same reasoning for the controller.

* Move contoller.reset() to setupController

Resetting the controller state is typically done in the setupController hook instead of in the willTransition action.

* Rename dataIsGettingLost to showConfirmationDialog

* Rename immediateCancel to resetChanges

* Remove a useless async

* Rename gentleCancel to cancel

* Rename bewaar to save

* Rename annuleer to cancel

* Rename verwerp to reset

* Remove a pair of redundant parentheses

* Fix the value supplied as bestuursfunctie id

'params' is not accessible here because the model hook can not access the query params of a parent route.

* Add a missing class

This class was mistakenly removed in a previous step.

* Remove redundant legacy condition

* Show correct bestuursfunctie.rol in the page title

* Use RSVP.hash instead of a POJO.

* Remove indentations on blank lines

* Disable UI buttons while saving

* Set bestuurseenheid in the controller

* Add a space between two words

* Fix a typo

* Remove pointless documentation

* Move reset logic to its appropriate action/method

* Move reset logic to its appropriate action/method

* Move reset logic to its appropriate action/method

* Replace classBinding with diabled

* Get rid of two redundant properties

* Use the functionaris-cloase-confirmation component

* Add a new computed property: isDirty

* Replcae 'Ember.computed' with 'computed'

* Remove unused argument

* Syntax fixes

* Fix missing action on functionaris table

* Simplify functionaris edit controller

* Simplify functionaris form

* Merge routes to create a new functionaris in 1 route

* Cleanup router paths

* Allow to clear adres of bestuursfunctie contact info