Handle case where input is not nice or roundable
When an input is not nice, or when its sum does not yield a nice
value, we should be rendering a number that comes close to the value
input by the user.  This is what we do here, coping with inputs from
textfields when the user's locale is not English.

Implementation is not perfect and could be incorrect when the numbers
pass the 10k mark.  I doubt this will happen and don't see a reason to
overengineer this feature.

A polyfill may be needed when running this on IE.