Remove the ember-cli-mirage testing setup (#102)
The setup was no longer working and getting it to work again is very time consuming for little benefit. The tests were alle failling and outdated as well.

Since the testing team is working towards full end-to-end testing it's probably better to simply drop the setup. We can always add a new modern setup in the future if needed.