Explicitly return the model data
When not defining a model hook, Ember will return the model value of the first parent route that defines one. This behaviour isn't really documented and is confusing when reading the code. Explicitly returning the model data "documents" where we expect our data to come from and gives us a pointer where to look if something goes wrong.