Convert dates in timezones with date-fns-tz
When selecting the start or end date of a mandate, the datepicker will
always choose midnight in the timezone setting of the user. This is not
correct, because we want to specify a date in the timezone of the
administrative unit, namely Belgium. Using date-fns-(tz) we can convert
and offset the date easily.