Move the contact editing to the mandataris edit page
This reworks the flow so we don't need separate contact point edit and creating pages. This makes the flow more consistent with the OrganizationPortal app.

It also solves the issue where rolling back changes, and updating contact point changes without affecting the mandataris record wasn't possible in the previous setup.