Always show the projectName if it exists
It's possible that we will use this field in other subsidy types in the future.

The current implementation assumes that all forms that belong to a certain subsidy application use the same project name. At the moment this isn't always the case because the shared form data is only synced when the forms are requested, but fixing that requires more architectural changes.