Prefill the mandatee endDate from the expectedEndDate
This was made for data quality, this way, every time the user picks a mandate it will autofill the endDate from the expectedEndDate. User has also the choice to input a date manually, it will stay even if the mandate changes but he will be warn with a warning message.