Fix an issue that occurred going to the subsidy details page
The details page requires some extra relationships to be loaded. If you start from the overview page, the base record is already loaded and Ember Data will return it from the cache. Some code depends on this data being loaded in the model hooks before being rendered. We now force a reload so that the model hooks properly await the full data set.