Update to Appuniversum 2.5 (#302)
One of the forms in "toezicht" uses a checkbox component that requires Appuniversum 2.5. Our SemVer version  range should support that already, but it seems the npm overrides config made it ignore that requirement and we were still on 2.4. This meant that an error would be thrown because the component could not be found.

We can safely remove the override since we updated to the latest beta of ember-acmidm-login which no longer brings in a v1 version.