Remove the -prod image builds
The difference between the QA and PRD builds is very small but it requires more manual actions when deploying. We now convert this to a single image by retrieving the ACMIDM config from the server.

The test selectors are now no longer stripped in production builds, but that doesn't matter that much for the final build size.