Don't show the unsaved changes modal for read-only forms
In some cases, form fields update the form store even when the form is in read-only mode. This is essentially a bug in the fields, and it should be fixed, but it's easy to introduce such regressions.

This change simply removes the change handlers so that the modal will never be shown when the form is in read-only mode. That way the popup will not be displayed even if fields update the form store by accident.