Don't focus the rich-text editor window when setting the initial value
The `controller.setHtmlContent` method focuses the editor window by default. Since we use the `focusout` event to detect changes this also has the side-effect that the "unsaved changes" modal is sometimes displayed even if no changes were made. Not focusing the editor solves at least one of the cases why this happens.