Add the `convert-to-currency` helper
This helper is used in some of the custom subsidy form field components which were transfered from the `ember-submission-form-fields` addon. It seems we forgot to transfer this helper but it was still working since the addon still exported it. We move it here now so it can be removed from the addon.