[DL-5849] Refactor the impersonation feature (#384)
The backend service changed implementations to work around some issues
which requires us to update the logic in the frontend as well.

The backend now impersonates "transparently" which means the `/current`
endpoints of the login services now return the impersonated account
instead of the admin one. We then need to fetch the original admin
account if an impersonation is active.