Custom inflector rule improvements
The current setup is very old and is based on the idea that all models
will have a Dutch name. Nowadays we use Dutch names, so the setup is
counter-intuitive.

This removes the Dutch rules and adds irregular rules for the words that need it. This makes the list simpler.

We now also use EmberData's inflector utilities instead of the
ember-inflector ones, since those have been deprecated.