Update the linting setup to an eslint v9 based setup
This updates the linting setup to the one that will be part of the blueprint soon: https://github.com/ember-cli/ember-cli/pull/10516

This setup is fully compatible with .gjs, so we don't need a stopgap fix due to the eslint-plugin-ember upgrade.

We did have to make some modifications since eslint-plugin-ember doesn't
expose the flat configs as intended yet.