Enable route bundle splitting
We configure Embroider to split the js bundles for each of the major routes. Embroider will then bundle the shared code in the initial bundle and load the other bundles lazily, when the routes are first accessed. This speeds up the initial load of the app and users won't download code of modules they don't have access to (or simply don't access).