Update the way we enable the new loket styles
The style overrides cause issues with the built-in modules.
To reduce the problem we now only load the overrides styles on the new loket pages.

Ideally we clean up the overrides, so they don't cause these issues in the first place, but that can be done as a future improvement.