Attempt a more graceful app degradation if something is off with
favorites.

The previous version is a hard crash of the app, from where
it's unlikely users will be able to figure out how recover themselves.

There will still be people confused and blocked but at
least (probably) less people.

It's not perfect, the favorites tab is still broken and breaks the
app.